[#21] Move Most Movement and Tilt Zoom Functionality into Blueprint
I want to eliminate the `PF2BirdsEyeCameraPawnBase` class, since there's
an awkward split in functionality between math and callbacks done in the
Blueprint and those done in the C++ code.

This commit accomplishes these first few steps:
1. Movement is now done via a Floating Movement component in the
   Blueprint, rather than movement logic in the C++ base class.
2. The Tilt Zoom calculations are now all done in Blueprint and no
   longer exist in C++.
3. The parent class of the Bird's Eye camera pawn is no longer the
   `PF2BirdsEyeCameraPawnBase` class.

The only thing remaining to be ported before I can delete the base class
is logic related to zooming in and out and focusing on characters. That
logic is broken in the current implementation, so it won't be a straight
port like the others above were.