[#21] Implement Proper Zooming along Forward Axis of Bird's Eye Camera
This enables the player to zoom in and out along the angle that has been
established by tilt-zoom, allowing the tilt-zoom axis to provide a
coarse-grained zoom control while the zoom axis provides a fine-grained
zoom control. This makes it easier for the player to view as much or as
little of an encounter as they want.

Now, all that remains to be ported from the original base class is the
logic for focusing on actors and locations. I suspect that this can be
implemented in a cleaner way using line traces than the way that is
being used that relies on trig. assumptions.