[#21] Eliminate `UPF2BirdsEyeCameraPawnBase`
As mentioned in prior commits, this class is no longer used. Though the
methods inside might have some utility later in this project, we can
always resurrect and adapt the code from history or pull in and adapt
the original code from the UE4 RTS project by Nick Pruehs.

It's best to let the need for this functionality come up organically, to
avoid bloat and gold plating. I suspect we WILL need this at some point,
but by then we can just implement it in Blueprint.