[#21] Add Support for Limiting Camera from Roaming Too Far Away
This adds some new settings to the Bird's Eye Camera Pawn ("Camera
Pawn") that enables it to stop accepting movement input if it's straying
too far away from the character that the player is currently
controlling. In a larger game, this might need to sync-up with a fog of
war system. For now, though, this just prevents players from getting
lost in the map and/or gaining unwanted intel from elsewhere in the map
during an encounter.

One remaining bug: Somehow the changes in this commit have broken the
ability to interpolate the camera from third-person exploration mode to
camera pawn encounter mode, but only for clients, not the listen server.
Not sure what this could be caused by, but replication is always a
suspect.