[#21] Add Callbacks for Changes to Character Speed + Include Current Val
This adds new native and BP-based callbacks that fire on the default
OpenPF2 attribute set and OpenPF2 interface, to allow the attribute
set and character to react to changes in the character's speed.

Along with this change, the current value of the changed attribute is
now passed through. The `OnHitPointsChanged` callback was modified in
the same way, for consistency.