[#8] Fix Obscure Linking Bug with `LoadBlueprint` Template in Specs
I am not sure why this worked up until now, nor am I sure why this issue
suddenly came up, but this should fix it. The issue would appear when
declaring a `UFUNCTION()` that takes a `TMap<int32,int32>` parameter.
Perhaps this is enough for UHT to split the code into separate
compilation units, revealing the linker issue?

This was the error before this fix was applied:
```
  unresolved external symbol "protected: static class TSubclassOf<class UPF2AbilityBoostCalculationBase> __cdecl FPF2SpecBase::LoadBlueprint<class UPF2AbilityBoostCalculationBase>(class FString,class FString)" (??$LoadBlueprint@VUPF2AbilityBoostCalculationBase@@@FPF2SpecBase@@KA?AV?$TSubclassOf@VUPF2AbilityBoostCalculationBase@@@@VFString@@0@Z) referenced in function "protected: void __cdecl FPF2AbilityBoostSpec::LoadMMCs(void)" (?LoadMMCs@FPF2AbilityBoostSpec@@IEAAXXZ)

  UnrealEditor-OpenPF2Core-Win64-DebugGame.dll: [LNK1120] 1 unresolved externals
```

The thing is, this should never have compiled because templates have to
be fully declared in headers.