[#8] Implement a Custom, Generic "Invert" Blueprint Node for TMaps
This finally adds the ability to easily invert any type of map in a
blueprint, as long as that map has a value type that can be used as a
key (the type must implement `GetTypeHash()`).

This functionality is too complex to express through normal metadata on
a `UFUNCTION` because the key and value types on the return value are
expected to be the inverse of the key and value types of the input
value. To accomplish that feat, a custom K2Node was implemented that
initially appears with a wildcard input and output pin. Once the input
pin has been attached to a non-wildcard map input, the output will
automatically assume the inverse of the input type.

At blueprint compilation time, the custom K2Node is expanded to a
hidden, internal-only node from the new `PF2MapLibrary` blueprint
function library. The internal node is defined with a custom thunk and
accepts a wildcard input and wildcard map output. Blueprint would never
allow this node to be used directly since the output pin provides no
type information to later nodes that receive input from it. Instead, the
custom K2Node forces the types of the pins on the internal node to match
the types on its own pins. For example, if the K2Node received an input
of type "const Map<String, Int>" then the internal node will be wired up
to accept "const Map<String, Int>" and return "const Map<Int, String>".

Since the internal node accepts wildcard parameters, it has a custom
thunk to extract the input and output map references off the stack. The
implementation then calls the `GenericMap_Invert` native code to perform
the actual invert operation by swapping the pointers of every pair from
the source map.

This is a LOT of code for what should be a straightforward use case, but
it is understandable given that the type safety built into Blueprint is
there for a reason and can't reasonably be expected to support this use
case easily since the types on the map are changing radically during the
call and are generic to begin win.

The only known limitation is that the constness of map values is not
preserved during the inversion. This is a side effect of the fact that
blueprint doesn't have a way to track the constness of map keys; it only
tracks the constness of map values and the map itself. So, after
inversion, we will lose the constness of the map values (which are now
the keys).