[#8] Add `GetAbilities()` to OpenPF2 ASC Interface
I struggled with a good name for this method. We can't call it
`GetActivatableAbilities()` because that would shadow a non-virtual
method in the ASC base class. I wanted to call it
`GetActivatableOpenPf2Abilities()` or
`GetActivatableAbilityInterfaces()` but those names seem way too
verbose. So, for now, this seems like an acceptable name.