[#8] Stop Setting "Input Mode UI Only" for Dialog Boxes
This interferes with the ability for key bindings to be released while
in a dialog box. So, for example, if there is an ability select dialog
that should only appear while the left trigger or SHIFT key is held
down, it wasn't possible to have it dismissed upon releasing the
trigger/key.