[#8] WIP Editor Support Module for Enhancing DX for Intf. Events Objects
The new pattern of having UInterfaces expose an "events" object for
binding delegates through an interface is great for avoiding coupling to
a concrete implementation of an interface, but it adds friction when
trying to wire-up delegates while in a Blueprint that instantiates an
interface implementation (i.e., if you are in a `PF2PlayerController`
blueprint and the controller has a `PF2CharacterQueueComponent`, it
already knows the concrete type of the component and should be able to
place the event bindings directly in the Blueprint without having to
bind delegates through the `GetEvents()` method).

To fix this, a new editor support module has been added that will use
the `PF2EventsInterface` to reflect on what events an interface object
is exposing so that bindings can be added to those delegates as if they
were bound on the interface object itself rather than its events object.

Right now, events appear correctly in the "Details" panel of the editor,
but you cannot yet bind them in the Blueprint (the + button does nothing
yet).