[#8] [UE-78453] Eliminate Multicast Refresh of ASC Ability Owner
For multiplayer games, this multicast event seemed to provide clients
with a way to workaround UE-78453 in UE 4.27, but this no longer works
in UE 5.1. Instead, the recommended workaround is for the player
controller of the project to override the `SetPawn()` method and provide
a timer callback on the character being unpossessed to update its ASC
during the next frame after controller ownership has replicated.

For more details on the bug, see:
https://github.com/OpenPF2/UE-78453-repro

This workaround wasn't done in the core project because not every
project will be multiplayer and AFAIK it's not possible to set a timer
tick callback on the UFUNCTION of a `UInterface`, so we'd need to couple
to the `PF2CharacterBase` class rather than the generic interface. You
can see this workaround in action inside the OpenPF2Playground project.