[#39] Codify New Convention of Passing Event Emitter as First Event Arg.
If this carries a performance penalty, I may regret this later, but for
now this makes it much easier to implement loose coupling among event
emitters and objects bound to their delegates by not having the object
responding to an event have to keep track of what they bound to.
Instead, the emitter gets supplied directly to the event.