[#45] Update Root Coll. Delegate Component for UE 5.2 + FIX COLLISION!!
UE now exposes in `PrimitiveComponent.h` some of the internal
definitions we previously had to borrow from `PrimitiveComponent.cpp`.
In addition, it is much smarter to keep our copy of the CVars in a
different namespace than the engine, and to load them lazily to avoid
a race condition and/or fetching the CVars before they've been
initialized.

Finally -- and most importantly -- I caught a few typos that I believe
were causing the collision delegate component to get stuck/embedded in
geometry instead of properly applying collisions. With those typos
fixed, I am happy to report that the floating camera appears to be
colliding with the world properly!