[#47] Replace Explicit Action Points GE with Gameplay Events from MoPRS
Now, the encounter MoPRS sends gameplay events to the active character
when they are added or removed from an encounter and when their turns
start and end. This happens in the MoPRS rather than in the character
class so that the events don't get sent outside of encounters. This also
simplifies the AI task for starting turns since it now only is
responsible for notifying the MoPRS about a change in character turn.