[#47] Populate "Max Act. Points" and "Max React. Points" at Turn Start
I had thought about populating these values with other base character
stats, but the stats are specific to encounter mode so it didn't seem
like that was ideal. For now, we'll override the max on every turn. The
values are provided by a curve table, so a game designer can uniformly
override these values easily. If a game really needs separate control
over the max values, they can provide their own start-of-turn GA and GE.