[#47] Avoid Queuing Up Reset of Action Points on No-op Turns
Previously, if a character did not perform any action on their turn,
they'd still wind up with a reset of their action points queued up to
hit during their next turn which could give them the opportunity to
perform more actions during that next turn than they should be entitled
to.

For example, this was the issue:
1. Character takes no action during first turn.
2. Turn times out after 1 sec of inactivity.
3. GE for resetting action points gets queued to hit character in 6 sec.
4. No other character has actions queued yet.
5. Character gets another turn.
6. Character burns 2 action points in less than 6 secs.
7. Action points get reset.
8. Character can keep going for another 3 action points even though they
   should only have 1 action point left.

This solution isn't perfect. Perhaps it would be better if this was
handled via a cooldown GE.