[#47] Use GA Cooldowns to Enforce Six-second Turn Cooldown
This appears to eliminate quite a few edge cases from the prior
approach. Previously it was difficult to coordinate a reset of a
character's action points outside of their turn, but now we simply reset
points at the START of their turn, which is closer to PF2 rules.

The GA that's activated when a character is taking a turn uses a
six-second cooldown to prevent it from being activated again too soon.
That same GA is now what gives each character their action points at the
start of their turn (via a conditional instant GE that overrides the
points), so if they're on cooldown they don't get points to spend.

To prevent rapidly cycling among characters in the encounter, the AI
behavior for the encounter has been refactored so that it only times out
a character and switches to another character if there is any other
character who has actions queued. If no other character has actions
queued, then the character whose turn it is stays the active character
for however long it takes them to queue their first action. In addition,
cycling still follows initiative order, so any character along the way
from the previously active character to the character who has actions
queued will still get an opportunity to queue up and execute actions at
the appropriate spot for their initiative.