[#47] PRF: Avoid Calling into Game Mode for Adding/Removing Enc. Chars.
The game mode calls into the MoPRS to accomplish adding or removing
characters from encounters, so there's not a need for the MoPRS to call
into the game mode. That the game mode has methods for this is really
just provided as a convenience to code OUTSIDE the game mode to notify
the active MoPRS without needing a reference to it.

If we later discover that the Game Mode must be notified of changes to
the characters in encounters, we can make the MoPRS implement an event
emitter pattern for it that the game mode can subscribe to.