[#34] Allow Custom Construction of Ability Activation Payloads
This can be used by games to add additional information (e.g., targeting
information) to an ability that is being activated through a binding.
Even though a stock OpenPF2 player controller provides a method for
obtaining a target selection, I did not want the default behavior of the
binding component to be to inject this target into an ability being
activated because I want to leave the mechanics of target selection up
to the game designer.

Despite this, the OpenPF2Playground will provide an example of how to
override the stock Ability Bindings Component with a version that does
add target information to the payload of every bound ability during an
encounter.