[#34] Queue Commands when Blocked during Current Character Turn
Previously, we only queued a command for a character when it wasn't that
character's turn. Now, we also queue it if the command gets blocked (by
tags, for example). This may need more work if the reason why the
command was blocked is because of a temporary condition on the
character.