[#34] Introduce `AttemptToQueueCommand` into Commands/GM/MoPRS
This makes it possible to slip a command into the queue for a character
without first trying to execute it, which is necessary when converting
a ranged combat attack ability into invocation of a movement ability
followed by the attack ability.

This rolls back the change made in the previous commit, since it's not
ideal to queue blocked commands since they might be blocked for reasons
other than the character being out of action points.