[#34] Drop Last Element, Not New Element, during `EnqueueAt()`
When a size limit is in place, we want to prevent too many items being
at the end of the queue, not at the beginning or middle of the queue.
Without this, a movement ability added by an attack ability could get
dropped, resulting in an infinite loop with no attack actually taking
place.