[#53] Implement Array-optimized Version of `Reduce` Array Util. Method
The prior implementation produces a lot of needless copies of the result
array. Granted, this new implementation is not pure functional because
we are now introducing a side effect for the passed-in first argument,
but this is a pretty minimal tradeoff for better performance and a
shorter amount of code at call time.

This does not affect other uses of `Reduce` (e.g., for integers, maps,
etc.), but I may add an optimized version for maps later.