[#53] Do Not Send Contents of Inventory in "Inventory Changed" Events
Inventory could be quite large (e.g., hundreds of items), so it doesn't
make sense for us to lag networking to send all items out every time
inventory changes. There is also no way to know if a listener of this
event even wants the contents of inventory. So, for all these reasons,
let's not send inventory during updates.