[#53] Define `MakeOutgoingGameplayEffectSpecForCauser` and `...Weapon`
The default methods for creating a GE spec involve either setting the
effect causer to the avatar actor (which isn't the effect we want), or
cloning an existing GE spec and then modifying the effect causer (which
is not efficient).

These new methods make it much easier to create a GE spec where the
effect causer is the weapon inflicting the damage.