[#53] Clarify Why We Take the Highest Weapon Proficiency
The Core Rulebook doesn't seem to explicitly say how to handle multiple
*weapon* proficiencies that apply to the same weapon, but it does call
the proficiency a bonus and is pretty clear about how other bonuses of
the same type get handled (TL;DR they don't stack). So, this is about as
clear a citation as I can find for this logic.