[#53] Add Additional Methods for GE Containers with Weapons as Causer
This enables GEs to interrogate weapon data during damage execution. A
different approach would be for the weapons to instantiate their damage
GEs and ingest the necessary context into the GE via transient
aggregators ("temporary variables"), but that move a lot of the
responsibility of knowing what data the GE exec will need into the
weapon itself, which feels like the wrong approach.

Let's see how this shakes out.