[#53] Give Weapons a `OnGameplayEffectsContainerSpecGenerated` Callback
In addition to the static list of GEs in the GE container property of
weapons, this enables weapons to generate dynamic GE specs in order to
populate "set by caller" magnitude variables/temporary variables
(a.k.a., transient aggregators). This can be used by weapons that need
to apply additional damage, such as from runes, magical abilities, and
more.