[#53] Re-factor Weapons to Support Both SOURCE and TARGET GE Containers
This enables weapons to apply GEs to the attacker that wields them
during an attack. By doing this, we can now do all the attack and damage
calculations in a GE applied to the source of an attack and then apply
the result to the target in a separate GE applied to the target. This
avoids having to manipulate the source ASC directly in GE executions.