[#53] Clean-up `GetEffectInstigator` and Remove FIXME about Dmg. Source
I'm not sure exactly what I was going on about here. The instigator is
the character that initiated the attack, while the damage source is the
object/weapon that inflicted damage on the target. The two don't have to
be determined in any particular order, and they can be the same actor in
the case of a melee attack.