[#53] Eliminate `PF2GameplayAbilityUtilities::GetEffectInstigator()`
The more I looked at it, the more I realized that this method doesn't
really have any relevance in this project since the controlled pawn
could easily be a floating camera pawn or a different character in the
party. The instigator set in the GE should be more than sufficient for
this.