[#53] Emit Gameplay Event when Hit Points Change
This makes it easier for a condition check GA to respond to a change in
hit points. I could have implemented this using the Blueprint Event that
already exists on PF2 characters, but that would require many additional
graph nodes in Blueprint to convert this back into GAS callbacks/GEs and
might lack much of the context from the original GE activation. I
believe the current approach should be more efficient and won't require
designers to use as much boilerplate code in character blueprints.