[#53] Expand Adv. Add'l Effects GE Comp. to Support Target Checks
This makes it possible to base the application of conditional GEs on the
tags of the target, rather than just the source. This was possible via
application tag requirements of each GE, but that would require creation
of a lot more wrapper GEs that had the conditions and would, in turn,
have additional conditional GEs.