[#53] Fix Emission of Attr. Set Events when Lacking Instigator or Target
Previously, the code assumed that both an instigator and target were
available whenever a character's health changed. The health change could
come from other effects on the same character, in which case they could
lack an instigator or target because both are implied from the fact that
the GE is being applied to the owning character. The change in HP could
even occur entirely outside of a GA or GE activation.