[#53] First Working Proof of Concept of `BeforeAll` / Drop `AfterAll`
It does not seem possible to implement `AfterAll` because the automation
test system parcels out tests one by one to test runners so no runner
has enough context to know when all the tests in a given scope have run.

In addition, the approach of using a modified `AutomationTest.h` won't
work because it prevents us from including any header file that includes
the original version of the header. The Unreal build process combines
multiple CPP files that have the same headers into one CPP file, so if
two CPP files have different versions of the same header, the build will
fail because the compiler detects a redefinition of the automation
classes.