[#53] Add Support for a MoPRS to React to Unconsciousness + Death States
Previously, the MoPRS could only react to starting to die and recovering
from death. Now, the MoPRS can react to a character being knocked out
unconscious and recovering from being unconscious in a distinct way from
starting to die, AND the MoPRS can officially react to the character
dying as well.