[#53] Pass Along Instanced Ability when Assembling Outgoing GE Context
Knock-out and condition check GA logic requires instanced abilities so
that we have access to source ability information when calculating stats
like a character's dying score or whether they are dying at all (for a
non-lethal attack). Given this, when building a GE spec on the server
side, rather than passing the ability spec handle and owning actor info,
we actually need to pass the ability instance itself along. Otherwise,
we can only get the CDO for the ability, resulting in a loss of source
captures by the time we're into condition GA.

This might be heavier-weight and require some additional tweaking for
multiplayer support. Time (and profiling) will tell.