[#53] Rework Dying Cond. Tests to Use `Let()` and Minimize `BeforeAll()`
`BeforeAll()` is great except that the Session Frontend executes tests
in depth-first, alphabetic order, so if we use `BeforeAll()` to setup
default values at a higher level and override them inside nested scopes,
tests that run at a higher scope after a nested scope will end up with
the wrong values.

`Let()` provides a much more elegant solution to this problem, without
any requirement on the execution order of the tests.