[#53] Make Melee Strike GA Generic and Move from Playground into Core
The GA now has to be sub-classed so that a character-specific montage
can be supplied. Meanwhile, the damage events are now dispatched in
response to a Gameplay Event tag that should typically get emitted by an
Anim. Notify State, rather than hard-coding the event to happen only at
the end of the montage. This gives the montage designer direct control
over the precise moment that damage should register.

This approach was inspired by Epic's "Action RPG" sample project.