[#62] Fix Consistency of Public Includes
Any type that appears in a public header that is from a dependency
should necessitate that dependency being made into a public dependency.
There are likely a few other cases of this that I'm missing that "just
work" even thought they're wrong, but this is a start. I wish there was
a tool to detect this.