[#62] Switch to Enh. Specs. Plug in + Move Tests to `OpenPF2Tests` Mod.
Now, tests for the project live in a separate module that is only built
for uncooked builds. This is necessary because cooked builds appear to
not have automation test support (which makes some intuitive sense,
since automation tests would bloat the final product for no benefit
other than smoke testing).